Add an option to suppress non-essential output.
This commit adds a new writer backend that only outputs essential
information, so that abraxas can be more easily used from scripts.  For
example, if the user requests a password with this backend, the password
will simply be written to stdout.  No label will be displayed and no
fancy timeout will exist.  For example:

$ abraxas -Pq <account>
<password>
$ abraxas -Nq <account>
<username>

Note that I commandeered the '-q' flag (meaning 'quiet') for this
backend.  This flag previously meant 'question' and printed the security
questions for the given account.  The '-Q' flag now performs this
function.  I think this configuration makes sense, since '-q'
traditionally means 'quiet' and most of the "choose what to display"
options are uppercase already.